+++ PhotosApp/PhotosApp/ViewController.swift
@@ -7,43 +7,18 @@
 //
 
 import UIKit
-import Photos
 
-class ViewController: UIViewController, UICollectionViewDataSource, UICollectionViewDelegate {
+class ViewController: UIViewController{
     @IBOutlet weak var collectionView: UICollectionView!
-    
-    var allPhotos: PHFetchResult<PHAsset>?
-    let imageManager = PHCachingImageManager()
+
+    let dataSource = PhotosCollectionDataSource()
+    let delegate = PhotosCollectionDelegate()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        allPhotos = PHAsset.fetchAssets(with: nil)
-        collectionView.reloadData()
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell: UICollectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: "collectionViewCell", for: indexPath)
-        
-        
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
-        let options = PHImageRequestOptions()
-        options.isNetworkAccessAllowed = true
-        options.isSynchronous = false
-        imageManager.requestImage(for: (allPhotos?[indexPath.row])!, targetSize: CGSize(width: 100, height: 100), contentMode: PHImageContentMode.aspectFill, options: options) { (image, info) -> Void in
-            if (image != nil) {
-                (cell as! CollectionViewCell).image.image = image
-            }
-        }
+        self.collectionView.dataSource = self.dataSource
+        self.collectionView.delegate = self.delegate
+        self.collectionView.reloadData()
     }
     
 }
